# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Aced Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/assets/{id}/download-url:
    post:
      operationId: getAssetDownloadUrl
      tags: [library]
      summary: Get short-lived signed download URL for an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/creator/applications:
    post:
      operationId: submitCreatorApplication
      tags: [creators]
      summary: Submit creator application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatorApplicationRequest"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfileResponse"
  /v1/creator/verifications/upload-url:
    post:
      operationId: getVerificationUploadUrl
      tags: [creators]
      summary: Get signed URL for verification document upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
...
  /v1/admin/applications:
    get:
      operationId: getAdminApplications
      tags: [admin]
      summary: Get creator applications queue
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationsResponse"
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        traceId:
          type: string
...
    DownloadUrlResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            url:
              type: string
            expiresInSeconds:
              type: integer
            fileName:
              type: string
...
    CreatorApplicationRequest:
      type: object
      required: [headline, universityId, courseId, graduationYear, academicResult, agreedToTerms]
      properties:
        headline:
          type: string
        universityId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        graduationYear:
          type: integer
        academicResult:
          type: string
        agreedToTerms:
          type: boolean
    CreatorProfileResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
    UploadUrlRequest:
      type: object
      required: [fileName, mimeType, sizeBytes, claimType]
      properties:
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        claimType:
          type: string
    UploadUrlResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            uploadUrl:
              type: string
            storageKey:
              type: string
            expiresInSeconds:
              type: integer
...
    ApplicationsResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            type: object
...
  responses:
...
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...